Fix extension icon not triggering analysis when state is idle
When clicking the extension icon, if the analysis state was 'idle'
(e.g., after navigating back to a previous page), the sidebar would
open but no analysis would be triggered, leaving the user with no results.

This fix adds an else clause to trigger a new analysis when the state
is idle. Since the backend has caching, this will be fast for recently
analyzed products.